+++ frontend/src/components/AppSelectMulti.vue
@@ -146,8 +146,8 @@ export type Options = Option[];
 <script setup lang="ts">
 import { computed, nextTick, ref, watch } from "vue";
 import { isEqual, uniqueId } from "lodash";
-import { useFloating } from "@/util/composables";
 import { wrap } from "@/util/math";
+import { useFloating } from "@/util/use-floating";
 
 type Props = {
   /** two-way bound selected items state */
