+++ frontend/src/pages/explore/TabPhenotypeExplorer.vue
@@ -258,8 +258,8 @@ import ThePhenogrid from "@/components/ThePhenogrid.vue";
 import { snackbar } from "@/components/TheSnackbar.vue";
 import { appendToBody } from "@/global/tooltip";
 import { scrollTo } from "@/router";
-import { useQuery } from "@/util/composables";
 import { parse } from "@/util/object";
+import { useQuery } from "@/util/use-query";
 import examples from "./phenotype-explorer.json";
 
 /** common tooltip explaining how to use multi-select component */
