+++ frontend/src/pages/node/AssociationsTable.vue
@@ -111,7 +111,7 @@ import AppTable from "@/components/AppTable.vue";
 import type { Cols, Sort } from "@/components/AppTable.vue";
 import { snackbar } from "@/components/TheSnackbar.vue";
 import { getBreadcrumbs } from "@/pages/node/AssociationsSummary.vue";
-import { useQuery } from "@/util/composables";
+import { useQuery } from "@/util/use-query";
 
 type Props = {
   /** current node */
